+++ client/game/Empire/Index.js
@@ -7,8 +7,9 @@ import Vue from "vue";
 import VueI18n from "vue-i18n";
 import Vuelidate from "vuelidate";
 import VModal from "vue-js-modal";
-import store from "../store/store";
+import store from "./store";
 import Empire from "./Empire.vue";
+import cfg from "Config";
 import {getAreaMessages} from "../handlers/messages";
 import {getLocale, getAreaSlug, getMessagesVersion} from "../handlers/gameConstants";
 
@@ -17,6 +18,7 @@ if (document.getElementById("gameRoot")) {
     Vue.use(VueI18n);
     Vue.use(Vuelidate);
     Vue.use(VModal, { dialog: true });
+    cfg.DEBUG && console.log(`bootstrapping VueJS, locale ${locale}.`);
     getAreaMessages(locale, getAreaSlug(), getMessagesVersion(), messages => {
         let i18nConfig = {
             locale: locale,